+++ config/config.default.php
@@ -263,7 +263,13 @@ if (is_dir(__DIR__ . '/../../../../config')) {
     // Someone has installed Imbo via a custom composer.json, so the custom config is outside of
     // the vendor dir. Loop through all available php files in the config dir
     foreach (glob(__DIR__ . '/../../../../config/*.php') as $file) {
-        $config = array_replace_recursive($config, require $file);
+        $extraConfig = require $file;
+
+        if (!is_array($extraConfig)) {
+            continue;
+        }
+
+        $config = array_replace_recursive($config, $extraConfig);
     }
 } else if (file_exists(__DIR__ . '/config.php')) {
     $config = array_replace_recursive($config, require __DIR__ . '/config.php');